+++ docs/src/examples/components/Popup/Types/PopupExample.shorthand.tsx
@@ -1,8 +1,26 @@
 import React from 'react'
-import { Button, Popup } from '@stardust-ui/react'
+import { Button, Input, Popup } from '@stardust-ui/react'
 
-const PopupExample = () => (
-  <Popup trigger={<Button icon="expand" />} content="Add users to your feed." />
-)
+class PopupExample extends React.Component<any, any> {
+  state = { popupOpen: false }
+
+  togglePopup() {
+    this.setState(prev => ({ popupOpen: !prev.popupOpen }))
+  }
+
+  render() {
+    return (
+      <Popup
+        open={this.state.popupOpen}
+        onOpenChange={(e, newProps) => {
+          alert(`Popup is requested to change its open state to "${newProps.open}".`)
+          this.setState({ popupOpen: newProps.open })
+        }}
+        trigger={<Button icon="expand" onClick={() => this.togglePopup()} />}
+        content={<Input icon="search" placeholder="Search..." />}
+      />
+    )
+  }
+}
 
 export default PopupExample