+++ src/components/List/List.tsx
@@ -32,9 +32,13 @@ class List extends UIComponent<Extendable<IListProps>, any> {
   static propTypes = {
     as: customPropTypes.as,
 
+    /**
+     *  Used to set content when using childrenApi - internal only
+     *  @docSiteIgnore
+     */
     children: PropTypes.node,
 
-    /** Additional classes. */
+    /** Additional CSS class name(s) to apply.  */
     className: PropTypes.string,
 
     /** Toggle debug mode */
@@ -55,7 +59,7 @@ class List extends UIComponent<Extendable<IListProps>, any> {
     /** Accessibility behavior if overridden by the user. */
     accessibility: PropTypes.func,
 
-    /** Custom styles to be applied for component. */
+    /** Additional CSS styles to apply to the component instance.  */
     styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
 
     /** Override for theme site variables to allow modifications of component styling via themes. */
@@ -66,20 +70,6 @@ class List extends UIComponent<Extendable<IListProps>, any> {
     as: 'ul',
     accessibility: ListBehavior as Accessibility,
   }
-
-  static handledProps = [
-    'accessibility',
-    'as',
-    'children',
-    'className',
-    'debug',
-    'items',
-    'selection',
-    'styles',
-    'truncateContent',
-    'truncateHeader',
-    'variables',
-  ]
 
   static Item = ListItem