+++ src/components/Popup/Popup.tsx
@@ -1,29 +1,43 @@
 import * as React from 'react'
+import { createPortal } from 'react-dom'
 import * as PropTypes from 'prop-types'
+import _ from 'lodash'
 import { Popper, PopperChildrenProps } from 'react-popper'
-import rtlCSSJS from 'rtl-css-js'
 
-import { childrenExist, customPropTypes, UIComponent, IRenderResultConfig } from '../../lib'
-import { ItemShorthand, Extendable, ReactChildren } from '../../../types/utils'
-import { ComponentVariablesInput, ComponentPartStyle } from '../../../types/theme'
-import Portal from '../Portal'
-import PopupContent from './PopupContent'
+import { childrenExist, AutoControlledComponent, IRenderResultConfig, isBrowser } from '../../lib'
+import {
+  ComponentEventHandler,
+  ItemShorthand,
+  Extendable,
+  ReactChildren,
+} from '../../../types/utils'
+
+import Ref from '../Ref'
 import computePopupPlacement, { Alignment, Position } from './positioningHelper'
+
+import PopupContent from './PopupContent'
+
+import { PopupBehavior } from '../../lib/accessibility'
+import {
+  Accessibility,
+  AccessibilityActionHandlers,
+  IAccessibilityBehavior,
+} from '../../lib/accessibility/interfaces'
 
 const POSITIONS: Position[] = ['above', 'below', 'before', 'after']
 const ALIGNMENTS: Alignment[] = ['top', 'bottom', 'start', 'end', 'center']
 
 export interface IPopupProps {
+  accessibility?: Accessibility
   align?: Alignment
-  as?: any
-  basic?: boolean
   children?: ReactChildren
   className?: string
   content?: ItemShorthand | ItemShorthand[]
+  defaultOpen?: boolean
+  open?: boolean
+  onOpenChange?: ComponentEventHandler<IPopupProps>
   position?: Position
   trigger?: JSX.Element
-  styles?: ComponentPartStyle
-  variables?: ComponentVariablesInput
 }
 
 export interface IPopupState {
@@ -35,7 +49,7 @@ export interface IPopupState {
  * @accessibility This is example usage of the accessibility tag.
  * This should be replaced with the actual description after the PR is merged
  */
-export default class Popup extends UIComponent<Extendable<IPopupProps>, IPopupState> {
+export default class Popup extends AutoControlledComponent<Extendable<IPopupProps>, IPopupState> {
   public static displayName = 'Popup'
 
   public static className = 'ui-popup'
@@ -43,23 +57,36 @@ export default class Popup extends UIComponent<Extendable<IPopupProps>, IPopupSt
   public static Content = PopupContent
 
   public static propTypes = {
+    /** Accessibility behavior if overridden by the user. */
+    accessibility: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
+
     /** Alignment for the popup. */
     align: PropTypes.oneOf(ALIGNMENTS),
 
-    /** An element type to render as (string or function). */
-    as: customPropTypes.as,
+    /**
+     *  Used to set content when using childrenApi - internal only
+     *  @docSiteIgnore
+     */
+    children: PropTypes.node,
 
-    /** Basic CSS styling for the popup. */
-    basic: PropTypes.bool,
-
-    /** The popup content (deprecated). */
-    children: customPropTypes.disallow(['children']),
-
-    /** Additional classes. */
+    /** Additional CSS class name(s) to apply.  */
     className: PropTypes.string,
 
     /** The popup content. */
-    content: customPropTypes.itemShorthand,
+    content: PropTypes.any,
+
+    /** Initial value for 'open'. */
+    defaultOpen: PropTypes.bool,
+
+    /** Defines whether popup is displayed. */
+    open: PropTypes.bool,
+
+    /**
+     * Event for request to change 'open' value.
+     * @param {SyntheticEvent} event - React's original SyntheticEvent.
+     * @param {object} data - All props and proposed value.
+     */
+    onOpenChange: PropTypes.func,
 
     /**
      * Position for the popup. Position has higher priority than align. If position is vertical ('above' | 'below')
@@ -71,58 +98,60 @@ export default class Popup extends UIComponent<Extendable<IPopupProps>, IPopupSt
 
     /** Element to be rendered in-place where the popup is defined. */
     trigger: PropTypes.node,
-
-    /** Custom styles to be applied for component. */
-    styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
-
-    /** Override for theme site variables to allow modifications of component styling via themes. */
-    variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
 
-  public static handledProps = [
-    'align',
-    'as',
-    'basic',
-    'children',
-    'className',
-    'content',
-    'position',
-    'styles',
-    'trigger',
-    'variables',
-  ]
-
-  public static defaultProps = {
-    as: Portal,
+  public static defaultProps: IPopupProps = {
+    accessibility: PopupBehavior,
     align: 'start',
     position: 'above',
+  }
+
+  public static autoControlledProps = ['open']
+
+  private static isBrowserContext = isBrowser()
+
+  protected actionHandlers: AccessibilityActionHandlers = {
+    toggle: e =>
+      _.invoke(this.props, 'onOpenChange', e, { ...this.props, ...{ open: !this.props.open } }),
+    closeAndFocusTrigger: e => {
+      if (this.props.onOpenChange) {
+        _.invoke(this.props, 'onOpenChange', e, { ...this.props, ...{ open: false } })
+        _.invoke(this.state.triggerRef, 'focus')
+      }
+    },
   }
 
   public state = { triggerRef: undefined }
 
   public renderComponent({
-    ElementType,
-    classes,
-    rest,
     rtl,
+    accessibility,
   }: IRenderResultConfig<IPopupProps>): React.ReactNode {
-    const { children, trigger } = this.props
+    const { children, trigger, open } = this.props
 
     return (
-      <ElementType
-        className={classes.root}
-        {...rest}
-        trigger={trigger}
-        triggerRef={this.handleTriggerRef}
-      >
-        {childrenExist(children) ? children : this.renderContent(rtl)}
-      </ElementType>
+      <>
+        <Ref
+          innerRef={domNode => {
+            this.setState({ triggerRef: domNode })
+          }}
+        >
+          {React.cloneElement(childrenExist(children) ? children : (trigger as any), {
+            ...accessibility.attributes.trigger,
+            ...accessibility.keyHandlers.trigger,
+          })}
+        </Ref>
+        {open &&
+          Popup.isBrowserContext &&
+          createPortal(this.renderPopupContent(rtl, accessibility), document.body)}
+      </>
     )
   }
 
-  private renderContent(rtl: boolean): JSX.Element {
+  private renderPopupContent(rtl: boolean, accessibility: IAccessibilityBehavior): JSX.Element {
     const { align, position } = this.props
     const triggerRef = this.state.triggerRef
+
     const placement = computePopupPlacement({ align, position, rtl })
 
     return (
@@ -130,22 +159,30 @@ export default class Popup extends UIComponent<Extendable<IPopupProps>, IPopupSt
         <Popper
           placement={placement}
           referenceElement={triggerRef}
-          children={this.renderPopperChildren.bind(this, rtl)}
+          children={this.renderPopperChildren.bind(this, rtl, accessibility)}
         />
       )
     )
   }
 
-  private renderPopperChildren = (rtl: boolean, { ref, style }: PopperChildrenProps) => {
-    const { basic, content } = this.props
-    const computedStyle = rtl ? rtlCSSJS(style) : style
+  private renderPopperChildren = (
+    rtl: boolean,
+    accessibility: IAccessibilityBehavior,
+    { ref, style: popupPlacementStyles }: PopperChildrenProps,
+  ) => {
+    const { content } = this.props
 
     return (
-      <Popup.Content innerRef={ref} basic={basic} {...rtl && { dir: 'rtl' }} styles={computedStyle}>
-        {content}
-      </Popup.Content>
+      <Ref innerRef={domElement => ref(domElement)}>
+        <Popup.Content
+          {...rtl && { dir: 'rtl' }}
+          style={popupPlacementStyles}
+          {...accessibility.attributes.popup}
+          {...accessibility.keyHandlers.popup}
+        >
+          {content}
+        </Popup.Content>
+      </Ref>
     )
   }
-
-  private handleTriggerRef = (triggerRef: HTMLElement) => this.setState({ triggerRef })
 }