+++ src/lib/renderComponent.tsx
@@ -31,6 +31,7 @@ import {
   IFocusZone,
   FocusZone as FabricFocusZone,
 } from './accessibility/FocusZone'
+import { FOCUSZONE_WRAP_ATTRIBUTE } from './accessibility/FocusZone/focusUtilities'
 
 export interface IRenderResultConfig<P> {
   ElementType: React.ReactType<P>
@@ -53,6 +54,7 @@ export interface IRenderConfig {
   props: IPropsWithVarsAndStyles
   state: IState
   actionHandlers: AccessibilityActionHandlers
+  focusZoneRef: (focusZone: IFocusZone) => void
 }
 
 const getAccessibility = (
@@ -88,8 +90,32 @@ function wrapInGenericFocusZone<
   FocusZone: { new (...args: any[]): COMPONENT },
   props: PROPS | undefined,
   children: React.ReactNode,
+  ref: (focusZone: IFocusZone) => void,
 ) {
-  return <FocusZone {...props}>{children}</FocusZone>
+  props[FOCUSZONE_WRAP_ATTRIBUTE] = true
+  return (
+    <FocusZone ref={ref} {...props}>
+      {children}
+    </FocusZone>
+  )
+}
+
+const renderWithFocusZone = (render, focusZoneDefinition, config, focusZoneRef): any => {
+  if (focusZoneDefinition.mode === FocusZoneMode.Wrap) {
+    return wrapInGenericFocusZone(
+      FabricFocusZone,
+      focusZoneDefinition.props,
+      render(config),
+      focusZoneRef,
+    )
+  }
+  if (focusZoneDefinition.mode === FocusZoneMode.Embed) {
+    const originalElementType = config.ElementType
+    config.ElementType = FabricFocusZone as any
+    config.rest = { ...config.rest, ...focusZoneDefinition.props }
+    config.rest.as = originalElementType
+  }
+  return render(config)
 }
 
 const renderComponent = <P extends {}>(
@@ -104,13 +130,14 @@ const renderComponent = <P extends {}>(
     props,
     state,
     actionHandlers,
+    focusZoneRef,
   } = config
 
   return (
     <FelaTheme
       render={(theme: IThemePrepared) => {
         const {
-          siteVariables = {},
+          siteVariables = { fontSizes: {} },
           componentVariables = {},
           componentStyles = {},
           rtl = false,
@@ -165,12 +192,10 @@ const renderComponent = <P extends {}>(
           theme,
         }
 
-        const rendered = render(config)
-
-        if (accessibility.focusZone && accessibility.focusZone.mode === FocusZoneMode.Wrap) {
-          return wrapInGenericFocusZone(FabricFocusZone, accessibility.focusZone.props, rendered)
+        if (accessibility.focusZone) {
+          return renderWithFocusZone(render, accessibility.focusZone, config, focusZoneRef)
         }
-        return rendered
+        return render(config)
       }}
     />
   )