+++ src/lib/typescriptUtils.ts
@@ -1,0 +1,5 @@
+const nameof = <T>(name: keyof T) => name
+
+export default {
+  nameof,
+}